Fix error handling in the pages and the function getUserId

+++ src/lib/actions/user.ts
@@ -14,20 +14,25 @@ export async function getAllUsers() {
 
 export const getUserById = async () => {
     const id = await getUserId();
-    const user = await db
-        .select({
-            userId: usersTable.userId,
-            fullName: usersTable.fullName,
-            age: usersTable.age,
-            email: usersTable.email,
-            imageUrl: usersTable.imageUrl,
-        })
-        .from(usersTable)
-        .where(eq(usersTable.userId, Number(id)));
-    if (!user) {
-        throw new Error("User not found");
+    if (!id) {
+        return null;
     }
-    return user;
+    try {
+        const user = await db
+            .select({
+                userId: usersTable.userId,
+                fullName: usersTable.fullName,
+                age: usersTable.age,
+                email: usersTable.email,
+                imageUrl: usersTable.imageUrl,
+            })
+            .from(usersTable)
+            .where(eq(usersTable.userId, Number(id)));
+        return user;
+    } catch (error) {
+        console.log(error);
+        return null;
+    }
 };
 
 export async function getUserId() {
@@ -42,7 +47,7 @@ export async function getUserId() {
         .where(eq(usersTable.clerkId, clerkId as string));
 
     if (user.length === 0) {
-        throw new Error("User not found");
+        return null;
     }
     const id = user[0].userId;
 
@@ -51,6 +56,10 @@ export async function getUserId() {
 
 export async function getCurrentUserInfo() {
     const userId = await getUserId();
+
+    if (!userId) {
+        return null;
+    }
 
     const user = await db
         .select({
@@ -142,6 +151,12 @@ export const updateUser = async (prevData: any, formData: FormData) => {
         };
     }
     const userId = await getUserId();
+    if (!userId) {
+        return {
+            success: false,
+            message: "Usuário não existe",
+        };
+    }
     const fullName = formData.get("fullName") as string;
     const email = formData.get("email") as string;
     const age = formData.get("age");
@@ -185,15 +200,20 @@ export const updateUser = async (prevData: any, formData: FormData) => {
 
 export const deleteUser = async () => {
     const userId = await getUserId();
+    if (!userId) {
+        return {
+            success: false,
+            message: "Usuário não existe",
+        };
+    }
     try {
         await db.delete(usersTable).where(eq(usersTable.userId, userId));
-        console.log("Passou...");
+
         return {
             success: true,
             message: "Perfil deletado com sucesso!",
         };
     } catch (error) {
-        console.log("Não passou");
         return {
             success: false,
             message: `Erro ao deletar perfil: ${error}`,
